refactor dcos-checks
